Minor refactoring: moved RecordReaderFactory to the factory package and updated classes that reference it.
